# DOCUMENTACION Y SCRPT/logs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SmarWatt Microservices API",
    "description": "API completa para análisis energético y chatbot inteligente",
    "version": "2.0.0",
    "contact": {
      "name": "SmarWatt Team",
      "email": "dev@smarwatt.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Energy IA API Development"
    },
    {
      "url": "http://localhost:5001",
      "description": "Expert Bot API Development"
    }
  ],
  "paths": {
...
    "/consumption": {
      "post": {
        "summary": "POST /consumption",
        "description": "Endpoint ULTRA-ROBUSTO para subir facturas de consumo.\n    NUNCA falla y SIEMPRE ofrece alternativas amables al usuario.\n    Garantiza máximo valor y experiencia perfecta.\n\n    IDÉNTICO AL ORIGINAL - PROHIBIDO CAMBIAR",
        "tags": [
          "Expert Bot API"
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error en la petición"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "message": {
                    "type": "string",
                    "example": "Mensaje de ejemplo"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/dashboard": {
      "get": {
        "summary": "GET /dashboard",
        "description": "Endpoint ULTRA-ROBUSTO para obtener datos del panel.\n    NUNCA falla y SIEMPRE proporciona datos útiles al usuario.\n\n    IDÉNTICO AL ORIGINAL - PROHIBIDO CAMBIAR",
        "tags": [
          "Expert Bot API"
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error en la petición"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      }
    },
    "/users/profile": {
      "get": {
        "summary": "GET /users/profile",
        "description": "Endpoint para obtener el perfil de usuario con datos de consumo.\n    Usado por energy_ia_api para obtener datos del usuario.\n\n    IDÉNTICO AL ORIGINAL - PROHIBIDO CAMBIAR",
        "tags": [
          "Expert Bot API"
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error en la petición"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      }
    },
    "/manual-data": {
      "post": {
        "summary": "POST /manual-data",
        "description": "NUEVO ENDPOINT REAL: Permite entrada manual cuando OCR falla o es incompleto.\n    Da VALOR INMEDIATO al usuario que no puede usar OCR por cualquier motivo.\n\n    IDÉNTICO AL ORIGINAL - PROHIBIDO CAMBIAR",
        "tags": [
          "Expert Bot API"
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error en la petición"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "message": {
                    "type": "string",
                    "example": "Mensaje de ejemplo"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/consumption/update": {
      "put": {
        "summary": "PUT /consumption/update",
        "description": "NUEVO ENDPOINT EMPRESARIAL: Actualizar datos de consumo existentes.\n    Permite actualizar datos sin subir nueva factura.",
        "tags": [
          "Expert Bot API"
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error en la petición"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "message": {
                    "type": "string",
                    "example": "Mensaje de ejemplo"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/consumption/history": {
      "get": {
        "summary": "GET /consumption/history",
        "description": "NUEVO ENDPOINT EMPRESARIAL: Obtener historial de consumos del usuario.\n    Funcionalidad empresarial para análisis de patrones.",
        "tags": [
          "Expert Bot API"
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error en la petición"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      }
    },
    "/consumption/analyze": {
      "post": {
        "summary": "POST /consumption/analyze",
        "description": "NUEVO ENDPOINT EMPRESARIAL: Análisis avanzado de consumo con IA.\n    Proporciona insights inteligentes sobre patrones de consumo.",
        "tags": [
          "Expert Bot API"
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error en la petición"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "message": {
                    "type": "string",
                    "example": "Mensaje de ejemplo"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/consumption/recommendations": {
      "get": {
        "summary": "GET /consumption/recommendations",
        "description": "NUEVO ENDPOINT EMPRESARIAL: Obtener recomendaciones personalizadas.\n    Basadas en IA y aprendizaje automático.",
        "tags": [
          "Expert Bot API"
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error en la petición"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token JWT de autenticación de Firebase"
      }
    }
  }
}